+++ pkg/kubelet/dockershim/network/OWNERS
@@ -1,0 +1,8 @@
+approvers:
+- thockin
+- dchen1107
+- matchstick
+- freehan
+reviewers:
+- sig-network-reviewers
+
