+++ pkg/kubelet/dockershim/network/cni/cni_test.go
@@ -39,10 +39,10 @@ import (
 	"k8s.io/kubernetes/pkg/kubelet/apis/kubeletconfig"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 	containertest "k8s.io/kubernetes/pkg/kubelet/container/testing"
-	"k8s.io/kubernetes/pkg/kubelet/network"
-	"k8s.io/kubernetes/pkg/kubelet/network/cni/testing"
-	"k8s.io/kubernetes/pkg/kubelet/network/hostport"
-	networktest "k8s.io/kubernetes/pkg/kubelet/network/testing"
+	"k8s.io/kubernetes/pkg/kubelet/dockershim/network"
+	"k8s.io/kubernetes/pkg/kubelet/dockershim/network/cni/testing"
+	"k8s.io/kubernetes/pkg/kubelet/dockershim/network/hostport"
+	networktest "k8s.io/kubernetes/pkg/kubelet/dockershim/network/testing"
 	"k8s.io/utils/exec"
 	fakeexec "k8s.io/utils/exec/testing"
 )
